requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
attrs==25.1.0
discord-protos==0.0.2
discord.py-self @ git+https://github.com/dolfies/discord.py-self.git@353924282a3d5832fb40844877b783caee65092b
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
protobuf==5.29.3
tzdata==2025.1
tzlocal==5.3
yarl==1.18.3
